Report queue time metric with the expected `qt` identifier (#102)
Our backend expects the metric identifiers to be reported from all our
libraries as "short" versions, but we were sending `queue_time` instead
here, which the backend is translating correctly but shouldn't have to.

The backend will continue to have legacy support for `queue_time` for a
while for current adapters out there, but this changes the Python adapter
going forward  to report metrics using `qt` to match our other adapters.

https://github.com/judoscale/judoscale-ruby/blob/ad739e4cf41e44a8e79c502afc02ffa461c5f8e1/judoscale-ruby/lib/judoscale/request_middleware.rb#L31
https://github.com/judoscale/judoscale-node/blob/408fe1db8e153b9f0d83cab3435b22d2e3a23e70/packages/express/src/index.js#L15
https://github.com/judoscale/judoscale-node/blob/408fe1db8e153b9f0d83cab3435b22d2e3a23e70/packages/fastify/src/plugin.js#L13