Fix running django/flask sample app tests
We need to set the runtime container configuration now, the previous
dyno config that was being set on these no longer exists.

These are apparently the only two sample apps with tests, and I want to
make sure they run successfully since we're adding them to the build as
well.