Fix fast API settings to not rely on pydantic base settings anymore
We don't need that behavior, and it's been extracted into a separate
dependency named `pydantic-settings` that only works with Python 3.9+
apparently, but we still support Python 3.8, so it's a whole dependency
conflict hell I do not want to mess with.

Other sample apps (including the other fast API sample with Celery)
simply uses a dict to setup these settings, so I went the same simpler
route here to get the app up and running again.

This also required changing the port to `5106` since npx is using `5006`
and we need to access the app through that now, so I updated places that
referenced the old `5000` port.