chore: Run tests from sample apps with the build (#103)
There are a couple sample web apps (flask, Django) with specific tests for them that weren't running as part of our build, this adds those tests to the build and fixes them.

I've also moved from request inspector to request catcher using the package name as subdomain now, it's the same monitor site we use on ruby/node, and I'll update them to use the same package name subdomain structure for consistency.