Fix calculating app time with default argument, convert to milliseconds
Python default arguments are immutable (which I sure had forgotten), so
we need to set the default in the method body instead to get a new value
every time.

I chose to still have that as a default so we don't need to be passing
it around from every caller, for simplicity there.

Also make sure we report app time as milliseconds to match queue time.