Decouple utilization from web collector, couple with metrics store
Instead of requiring the utilization tracker to be started with the web
collector, make the underlying metrics store a shared object we can
directly use by both, so we can push metrics directly to the underlying
store from the utilization tracker, and have the web metrics collector
flush them all together when reporting.

This allows us to remove the need to start the tracker from the init
middleware code, because that resulted in us starting a tracker for the
master process on our django sample app, which would track essentially
no processes handling requests every time because that master process
does not get requests sent to it.

Instead, we want to just start the utilization tracker the first time
the app gets hit by a request, in the middleware code itself, so it only
ever boots up on processes that actually handle requests, and we can
calculate utilization correctly, otherwise it could be skewed by
processes not handling requests bringing utilization average down.