Wrap the whole request middleware with utilization tracker
This is more inline with how Ruby is doing it, except there we have a
separate middleware but one that comes before the request middleware,
effectively wrapping it as well.

Also refactor it to use a context manager so we can use a `with` block
and have the code separated, without needing to implement additional
methods on the tracker itself.

(I actually made it work with the whole __enter__ and __exit__ methods,
but this is kinda simpler, again similar to Ruby, and keeps the code
together)

https://docs.python.org/3/library/contextlib.html#contextlib.contextmanager
https://www.geeksforgeeks.org/python/with-statement-in-python/