Accept JUDOSCALE_URL when we cannot identify the runtime container
This allows the package to be used by a custom platform integration by
simply setting the `JUDOSCALE_URL`, otherwise it would not initialize
since it couldn't identify the runtime container.

We might explore allowing a custom runtime container ENV var to be
passed that is not one of the specific platforms in the future.