fix: Accept JUDOSCALE_URL when we cannot identify the runtime container (#112)
This allows the package to be used by a custom platform integration by
simply setting the `JUDOSCALE_URL`, otherwise it would not initialize
since it couldn't identify the runtime container.

We might explore allowing a custom runtime container ENV var to be
passed that is not one of the specific platforms in the future.

Remove check for runtime container from reporter start.
Not having a container is not necessarily a blocker to start the
reporter. This makes the python package more inline with ruby & node.