Collect & report task metadata for Celery
This additional task metadata will help troubleshooting an issue with
some queue time spikes that we're investigating.

With each queue we'll add information about the task and the timestamps
that were used to calculate queue time, so it can hopefully help us
identify the source of the queue time spikes.