Test suggested ENV var to fix the problem for now
I'm running into some test trouble with supporting Python v3.8 all the
way to v3.14 now, and what seems to be incompatibilities here and there.

This fixes the last build issue to get v3.14 working:

    error: the configured Python interpreter version (3.14) is newer than PyO3's maximum supported version (3.13)
    = help: please check if an updated version of PyO3 is available. Current version: 0.22.2
    = help: set PYO3_USE_ABI3_FORWARD_COMPATIBILITY=1 to suppress this check and build anyway using the stable ABI

I couldn't find the magic incantation to get the fastapi tests running
with the right dependencies across all versions, so using this ENV var
for now to get it green.

https://github.com/PyO3/pyo3/blob/v0.27.1/CHANGELOG.md#0271---2025-10-21

I'll circle back on it as we will be dropping support to older versions
at some point soon hopefully, so for now I just want to get the build on
Python v3.14 going.